Mqtt_assignment_exalens/requirements.txt
paho-mqtt==1.5.1
pymongo==3.12.0
redis==3.5.3
fastapi==0.68.0
uvicorn==0.15.0
